Move in-between bitboard table to a dedicated module
