+++ src-tauri/src/eval/bitboards/in_between.rs
@@ -1,0 +1,4341 @@
+use super::Bitboard;
+
+pub const IN_BETWEEN_TABLE: [[Bitboard; 64]; 64] = [
+    [
+        0,
+        0,
+        0x0000000000000002,
+        0x0000000000000006,
+        0x000000000000000e,
+        0x000000000000001e,
+        0x000000000000003e,
+        0x000000000000007e,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000000100,
+        0,
+        0x0000000000000200,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000010100,
+        0,
+        0,
+        0x0000000000040200,
+        0,
+        0,
+        0,
+        0,
+        0x0000000001010100,
+        0,
+        0,
+        0,
+        0x0000000008040200,
+        0,
+        0,
+        0,
+        0x0000000101010100,
+        0,
+        0,
+        0,
+        0,
+        0x0000001008040200,
+        0,
+        0,
+        0x0000010101010100,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000201008040200,
+        0,
+        0x0001010101010100,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0040201008040200,
+    ],
+    [
+        0,
+        0,
+        0,
+        0x0000000000000004,
+        0x000000000000000c,
+        0x000000000000001c,
+        0x000000000000003c,
+        0x000000000000007c,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000000200,
+        0,
+        0x0000000000000400,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000020200,
+        0,
+        0,
+        0x0000000000080400,
+        0,
+        0,
+        0,
+        0,
+        0x0000000002020200,
+        0,
+        0,
+        0,
+        0x0000000010080400,
+        0,
+        0,
+        0,
+        0x0000000202020200,
+        0,
+        0,
+        0,
+        0,
+        0x0000002010080400,
+        0,
+        0,
+        0x0000020202020200,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000402010080400,
+        0,
+        0x0002020202020200,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+    ],
+    [
+        0x0000000000000002,
+        0,
+        0,
+        0,
+        0x0000000000000008,
+        0x0000000000000018,
+        0x0000000000000038,
+        0x0000000000000078,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000000200,
+        0,
+        0x0000000000000400,
+        0,
+        0x0000000000000800,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000040400,
+        0,
+        0,
+        0x0000000000100800,
+        0,
+        0,
+        0,
+        0,
+        0x0000000004040400,
+        0,
+        0,
+        0,
+        0x0000000020100800,
+        0,
+        0,
+        0,
+        0x0000000404040400,
+        0,
+        0,
+        0,
+        0,
+        0x0000004020100800,
+        0,
+        0,
+        0x0000040404040400,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0004040404040400,
+        0,
+        0,
+        0,
+        0,
+        0,
+    ],
+    [
+        0x0000000000000006,
+        0x0000000000000004,
+        0,
+        0,
+        0,
+        0x0000000000000010,
+        0x0000000000000030,
+        0x0000000000000070,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000000400,
+        0,
+        0x0000000000000800,
+        0,
+        0x0000000000001000,
+        0,
+        0,
+        0x0000000000020400,
+        0,
+        0,
+        0x0000000000080800,
+        0,
+        0,
+        0x0000000000201000,
+        0,
+        0,
+        0,
+        0,
+        0x0000000008080800,
+        0,
+        0,
+        0,
+        0x0000000040201000,
+        0,
+        0,
+        0,
+        0x0000000808080800,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000080808080800,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0008080808080800,
+        0,
+        0,
+        0,
+        0,
+    ],
+    [
+        0x000000000000000e,
+        0x000000000000000c,
+        0x0000000000000008,
+        0,
+        0,
+        0,
+        0x0000000000000020,
+        0x0000000000000060,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000000800,
+        0,
+        0x0000000000001000,
+        0,
+        0x0000000000002000,
+        0,
+        0,
+        0x0000000000040800,
+        0,
+        0,
+        0x0000000000101000,
+        0,
+        0,
+        0x0000000000402000,
+        0x0000000002040800,
+        0,
+        0,
+        0,
+        0x0000000010101000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000001010101000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000101010101000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0010101010101000,
+        0,
+        0,
+        0,
+    ],
+    [
+        0x000000000000001e,
+        0x000000000000001c,
+        0x0000000000000018,
+        0x0000000000000010,
+        0,
+        0,
+        0,
+        0x0000000000000040,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000001000,
+        0,
+        0x0000000000002000,
+        0,
+        0x0000000000004000,
+        0,
+        0,
+        0x0000000000081000,
+        0,
+        0,
+        0x0000000000202000,
+        0,
+        0,
+        0,
+        0x0000000004081000,
+        0,
+        0,
+        0,
+        0x0000000020202000,
+        0,
+        0,
+        0x0000000204081000,
+        0,
+        0,
+        0,
+        0,
+        0x0000002020202000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000202020202000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0020202020202000,
+        0,
+        0,
+    ],
+    [
+        0x000000000000003e,
+        0x000000000000003c,
+        0x0000000000000038,
+        0x0000000000000030,
+        0x0000000000000020,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000002000,
+        0,
+        0x0000000000004000,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000102000,
+        0,
+        0,
+        0x0000000000404000,
+        0,
+        0,
+        0,
+        0x0000000008102000,
+        0,
+        0,
+        0,
+        0x0000000040404000,
+        0,
+        0,
+        0x0000000408102000,
+        0,
+        0,
+        0,
+        0,
+        0x0000004040404000,
+        0,
+        0x0000020408102000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000404040404000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0040404040404000,
+        0,
+    ],
+    [
+        0x000000000000007e,
+        0x000000000000007c,
+        0x0000000000000078,
+        0x0000000000000070,
+        0x0000000000000060,
+        0x0000000000000040,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000004000,
+        0,
+        0x0000000000008000,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000204000,
+        0,
+        0,
+        0x0000000000808000,
+        0,
+        0,
+        0,
+        0x0000000010204000,
+        0,
+        0,
+        0,
+        0x0000000080808000,
+        0,
+        0,
+        0x0000000810204000,
+        0,
+        0,
+        0,
+        0,
+        0x0000008080808000,
+        0,
+        0x0000040810204000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000808080808000,
+        0x0002040810204000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0080808080808000,
+    ],
+    [
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000000200,
+        0x0000000000000600,
+        0x0000000000000e00,
+        0x0000000000001e00,
+        0x0000000000003e00,
+        0x0000000000007e00,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000010000,
+        0,
+        0x0000000000020000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000001010000,
+        0,
+        0,
+        0x0000000004020000,
+        0,
+        0,
+        0,
+        0,
+        0x0000000101010000,
+        0,
+        0,
+        0,
+        0x0000000804020000,
+        0,
+        0,
+        0,
+        0x0000010101010000,
+        0,
+        0,
+        0,
+        0,
+        0x0000100804020000,
+        0,
+        0,
+        0x0001010101010000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0020100804020000,
+        0,
+    ],
+    [
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000000400,
+        0x0000000000000c00,
+        0x0000000000001c00,
+        0x0000000000003c00,
+        0x0000000000007c00,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000020000,
+        0,
+        0x0000000000040000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000002020000,
+        0,
+        0,
+        0x0000000008040000,
+        0,
+        0,
+        0,
+        0,
+        0x0000000202020000,
+        0,
+        0,
+        0,
+        0x0000001008040000,
+        0,
+        0,
+        0,
+        0x0000020202020000,
+        0,
+        0,
+        0,
+        0,
+        0x0000201008040000,
+        0,
+        0,
+        0x0002020202020000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0040201008040000,
+    ],
+    [
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000000200,
+        0,
+        0,
+        0,
+        0x0000000000000800,
+        0x0000000000001800,
+        0x0000000000003800,
+        0x0000000000007800,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000020000,
+        0,
+        0x0000000000040000,
+        0,
+        0x0000000000080000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000004040000,
+        0,
+        0,
+        0x0000000010080000,
+        0,
+        0,
+        0,
+        0,
+        0x0000000404040000,
+        0,
+        0,
+        0,
+        0x0000002010080000,
+        0,
+        0,
+        0,
+        0x0000040404040000,
+        0,
+        0,
+        0,
+        0,
+        0x0000402010080000,
+        0,
+        0,
+        0x0004040404040000,
+        0,
+        0,
+        0,
+        0,
+        0,
+    ],
+    [
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000000600,
+        0x0000000000000400,
+        0,
+        0,
+        0,
+        0x0000000000001000,
+        0x0000000000003000,
+        0x0000000000007000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000040000,
+        0,
+        0x0000000000080000,
+        0,
+        0x0000000000100000,
+        0,
+        0,
+        0x0000000002040000,
+        0,
+        0,
+        0x0000000008080000,
+        0,
+        0,
+        0x0000000020100000,
+        0,
+        0,
+        0,
+        0,
+        0x0000000808080000,
+        0,
+        0,
+        0,
+        0x0000004020100000,
+        0,
+        0,
+        0,
+        0x0000080808080000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0008080808080000,
+        0,
+        0,
+        0,
+        0,
+    ],
+    [
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000000e00,
+        0x0000000000000c00,
+        0x0000000000000800,
+        0,
+        0,
+        0,
+        0x0000000000002000,
+        0x0000000000006000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000080000,
+        0,
+        0x0000000000100000,
+        0,
+        0x0000000000200000,
+        0,
+        0,
+        0x0000000004080000,
+        0,
+        0,
+        0x0000000010100000,
+        0,
+        0,
+        0x0000000040200000,
+        0x0000000204080000,
+        0,
+        0,
+        0,
+        0x0000001010100000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000101010100000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0010101010100000,
+        0,
+        0,
+        0,
+    ],
+    [
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000001e00,
+        0x0000000000001c00,
+        0x0000000000001800,
+        0x0000000000001000,
+        0,
+        0,
+        0,
+        0x0000000000004000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000100000,
+        0,
+        0x0000000000200000,
+        0,
+        0x0000000000400000,
+        0,
+        0,
+        0x0000000008100000,
+        0,
+        0,
+        0x0000000020200000,
+        0,
+        0,
+        0,
+        0x0000000408100000,
+        0,
+        0,
+        0,
+        0x0000002020200000,
+        0,
+        0,
+        0x0000020408100000,
+        0,
+        0,
+        0,
+        0,
+        0x0000202020200000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0020202020200000,
+        0,
+        0,
+    ],
+    [
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000003e00,
+        0x0000000000003c00,
+        0x0000000000003800,
+        0x0000000000003000,
+        0x0000000000002000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000200000,
+        0,
+        0x0000000000400000,
+        0,
+        0,
+        0,
+        0,
+        0x0000000010200000,
+        0,
+        0,
+        0x0000000040400000,
+        0,
+        0,
+        0,
+        0x0000000810200000,
+        0,
+        0,
+        0,
+        0x0000004040400000,
+        0,
+        0,
+        0x0000040810200000,
+        0,
+        0,
+        0,
+        0,
+        0x0000404040400000,
+        0,
+        0x0002040810200000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0040404040400000,
+        0,
+    ],
+    [
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000007e00,
+        0x0000000000007c00,
+        0x0000000000007800,
+        0x0000000000007000,
+        0x0000000000006000,
+        0x0000000000004000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000400000,
+        0,
+        0x0000000000800000,
+        0,
+        0,
+        0,
+        0,
+        0x0000000020400000,
+        0,
+        0,
+        0x0000000080800000,
+        0,
+        0,
+        0,
+        0x0000001020400000,
+        0,
+        0,
+        0,
+        0x0000008080800000,
+        0,
+        0,
+        0x0000081020400000,
+        0,
+        0,
+        0,
+        0,
+        0x0000808080800000,
+        0,
+        0x0004081020400000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0080808080800000,
+    ],
+    [
+        0x0000000000000100,
+        0,
+        0x0000000000000200,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000020000,
+        0x0000000000060000,
+        0x00000000000e0000,
+        0x00000000001e0000,
+        0x00000000003e0000,
+        0x00000000007e0000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000001000000,
+        0,
+        0x0000000002000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000101000000,
+        0,
+        0,
+        0x0000000402000000,
+        0,
+        0,
+        0,
+        0,
+        0x0000010101000000,
+        0,
+        0,
+        0,
+        0x0000080402000000,
+        0,
+        0,
+        0,
+        0x0001010101000000,
+        0,
+        0,
+        0,
+        0,
+        0x0010080402000000,
+        0,
+        0,
+    ],
+    [
+        0,
+        0x0000000000000200,
+        0,
+        0x0000000000000400,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000040000,
+        0x00000000000c0000,
+        0x00000000001c0000,
+        0x00000000003c0000,
+        0x00000000007c0000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000002000000,
+        0,
+        0x0000000004000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000202000000,
+        0,
+        0,
+        0x0000000804000000,
+        0,
+        0,
+        0,
+        0,
+        0x0000020202000000,
+        0,
+        0,
+        0,
+        0x0000100804000000,
+        0,
+        0,
+        0,
+        0x0002020202000000,
+        0,
+        0,
+        0,
+        0,
+        0x0020100804000000,
+        0,
+    ],
+    [
+        0x0000000000000200,
+        0,
+        0x0000000000000400,
+        0,
+        0x0000000000000800,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000020000,
+        0,
+        0,
+        0,
+        0x0000000000080000,
+        0x0000000000180000,
+        0x0000000000380000,
+        0x0000000000780000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000002000000,
+        0,
+        0x0000000004000000,
+        0,
+        0x0000000008000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000404000000,
+        0,
+        0,
+        0x0000001008000000,
+        0,
+        0,
+        0,
+        0,
+        0x0000040404000000,
+        0,
+        0,
+        0,
+        0x0000201008000000,
+        0,
+        0,
+        0,
+        0x0004040404000000,
+        0,
+        0,
+        0,
+        0,
+        0x0040201008000000,
+    ],
+    [
+        0,
+        0x0000000000000400,
+        0,
+        0x0000000000000800,
+        0,
+        0x0000000000001000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000060000,
+        0x0000000000040000,
+        0,
+        0,
+        0,
+        0x0000000000100000,
+        0x0000000000300000,
+        0x0000000000700000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000004000000,
+        0,
+        0x0000000008000000,
+        0,
+        0x0000000010000000,
+        0,
+        0,
+        0x0000000204000000,
+        0,
+        0,
+        0x0000000808000000,
+        0,
+        0,
+        0x0000002010000000,
+        0,
+        0,
+        0,
+        0,
+        0x0000080808000000,
+        0,
+        0,
+        0,
+        0x0000402010000000,
+        0,
+        0,
+        0,
+        0x0008080808000000,
+        0,
+        0,
+        0,
+        0,
+    ],
+    [
+        0,
+        0,
+        0x0000000000000800,
+        0,
+        0x0000000000001000,
+        0,
+        0x0000000000002000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x00000000000e0000,
+        0x00000000000c0000,
+        0x0000000000080000,
+        0,
+        0,
+        0,
+        0x0000000000200000,
+        0x0000000000600000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000008000000,
+        0,
+        0x0000000010000000,
+        0,
+        0x0000000020000000,
+        0,
+        0,
+        0x0000000408000000,
+        0,
+        0,
+        0x0000001010000000,
+        0,
+        0,
+        0x0000004020000000,
+        0x0000020408000000,
+        0,
+        0,
+        0,
+        0x0000101010000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0010101010000000,
+        0,
+        0,
+        0,
+    ],
+    [
+        0,
+        0,
+        0,
+        0x0000000000001000,
+        0,
+        0x0000000000002000,
+        0,
+        0x0000000000004000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x00000000001e0000,
+        0x00000000001c0000,
+        0x0000000000180000,
+        0x0000000000100000,
+        0,
+        0,
+        0,
+        0x0000000000400000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000010000000,
+        0,
+        0x0000000020000000,
+        0,
+        0x0000000040000000,
+        0,
+        0,
+        0x0000000810000000,
+        0,
+        0,
+        0x0000002020000000,
+        0,
+        0,
+        0,
+        0x0000040810000000,
+        0,
+        0,
+        0,
+        0x0000202020000000,
+        0,
+        0,
+        0x0002040810000000,
+        0,
+        0,
+        0,
+        0,
+        0x0020202020000000,
+        0,
+        0,
+    ],
+    [
+        0,
+        0,
+        0,
+        0,
+        0x0000000000002000,
+        0,
+        0x0000000000004000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x00000000003e0000,
+        0x00000000003c0000,
+        0x0000000000380000,
+        0x0000000000300000,
+        0x0000000000200000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000020000000,
+        0,
+        0x0000000040000000,
+        0,
+        0,
+        0,
+        0,
+        0x0000001020000000,
+        0,
+        0,
+        0x0000004040000000,
+        0,
+        0,
+        0,
+        0x0000081020000000,
+        0,
+        0,
+        0,
+        0x0000404040000000,
+        0,
+        0,
+        0x0004081020000000,
+        0,
+        0,
+        0,
+        0,
+        0x0040404040000000,
+        0,
+    ],
+    [
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000004000,
+        0,
+        0x0000000000008000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x00000000007e0000,
+        0x00000000007c0000,
+        0x0000000000780000,
+        0x0000000000700000,
+        0x0000000000600000,
+        0x0000000000400000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000040000000,
+        0,
+        0x0000000080000000,
+        0,
+        0,
+        0,
+        0,
+        0x0000002040000000,
+        0,
+        0,
+        0x0000008080000000,
+        0,
+        0,
+        0,
+        0x0000102040000000,
+        0,
+        0,
+        0,
+        0x0000808080000000,
+        0,
+        0,
+        0x0008102040000000,
+        0,
+        0,
+        0,
+        0,
+        0x0080808080000000,
+    ],
+    [
+        0x0000000000010100,
+        0,
+        0,
+        0x0000000000020400,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000010000,
+        0,
+        0x0000000000020000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000002000000,
+        0x0000000006000000,
+        0x000000000e000000,
+        0x000000001e000000,
+        0x000000003e000000,
+        0x000000007e000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000100000000,
+        0,
+        0x0000000200000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000010100000000,
+        0,
+        0,
+        0x0000040200000000,
+        0,
+        0,
+        0,
+        0,
+        0x0001010100000000,
+        0,
+        0,
+        0,
+        0x0008040200000000,
+        0,
+        0,
+        0,
+    ],
+    [
+        0,
+        0x0000000000020200,
+        0,
+        0,
+        0x0000000000040800,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000020000,
+        0,
+        0x0000000000040000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000004000000,
+        0x000000000c000000,
+        0x000000001c000000,
+        0x000000003c000000,
+        0x000000007c000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000200000000,
+        0,
+        0x0000000400000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000020200000000,
+        0,
+        0,
+        0x0000080400000000,
+        0,
+        0,
+        0,
+        0,
+        0x0002020200000000,
+        0,
+        0,
+        0,
+        0x0010080400000000,
+        0,
+        0,
+    ],
+    [
+        0,
+        0,
+        0x0000000000040400,
+        0,
+        0,
+        0x0000000000081000,
+        0,
+        0,
+        0x0000000000020000,
+        0,
+        0x0000000000040000,
+        0,
+        0x0000000000080000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000002000000,
+        0,
+        0,
+        0,
+        0x0000000008000000,
+        0x0000000018000000,
+        0x0000000038000000,
+        0x0000000078000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000200000000,
+        0,
+        0x0000000400000000,
+        0,
+        0x0000000800000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000040400000000,
+        0,
+        0,
+        0x0000100800000000,
+        0,
+        0,
+        0,
+        0,
+        0x0004040400000000,
+        0,
+        0,
+        0,
+        0x0020100800000000,
+        0,
+    ],
+    [
+        0x0000000000040200,
+        0,
+        0,
+        0x0000000000080800,
+        0,
+        0,
+        0x0000000000102000,
+        0,
+        0,
+        0x0000000000040000,
+        0,
+        0x0000000000080000,
+        0,
+        0x0000000000100000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000006000000,
+        0x0000000004000000,
+        0,
+        0,
+        0,
+        0x0000000010000000,
+        0x0000000030000000,
+        0x0000000070000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000400000000,
+        0,
+        0x0000000800000000,
+        0,
+        0x0000001000000000,
+        0,
+        0,
+        0x0000020400000000,
+        0,
+        0,
+        0x0000080800000000,
+        0,
+        0,
+        0x0000201000000000,
+        0,
+        0,
+        0,
+        0,
+        0x0008080800000000,
+        0,
+        0,
+        0,
+        0x0040201000000000,
+    ],
+    [
+        0,
+        0x0000000000080400,
+        0,
+        0,
+        0x0000000000101000,
+        0,
+        0,
+        0x0000000000204000,
+        0,
+        0,
+        0x0000000000080000,
+        0,
+        0x0000000000100000,
+        0,
+        0x0000000000200000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x000000000e000000,
+        0x000000000c000000,
+        0x0000000008000000,
+        0,
+        0,
+        0,
+        0x0000000020000000,
+        0x0000000060000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000800000000,
+        0,
+        0x0000001000000000,
+        0,
+        0x0000002000000000,
+        0,
+        0,
+        0x0000040800000000,
+        0,
+        0,
+        0x0000101000000000,
+        0,
+        0,
+        0x0000402000000000,
+        0x0002040800000000,
+        0,
+        0,
+        0,
+        0x0010101000000000,
+        0,
+        0,
+        0,
+    ],
+    [
+        0,
+        0,
+        0x0000000000100800,
+        0,
+        0,
+        0x0000000000202000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000100000,
+        0,
+        0x0000000000200000,
+        0,
+        0x0000000000400000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x000000001e000000,
+        0x000000001c000000,
+        0x0000000018000000,
+        0x0000000010000000,
+        0,
+        0,
+        0,
+        0x0000000040000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000001000000000,
+        0,
+        0x0000002000000000,
+        0,
+        0x0000004000000000,
+        0,
+        0,
+        0x0000081000000000,
+        0,
+        0,
+        0x0000202000000000,
+        0,
+        0,
+        0,
+        0x0004081000000000,
+        0,
+        0,
+        0,
+        0x0020202000000000,
+        0,
+        0,
+    ],
+    [
+        0,
+        0,
+        0,
+        0x0000000000201000,
+        0,
+        0,
+        0x0000000000404000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000200000,
+        0,
+        0x0000000000400000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x000000003e000000,
+        0x000000003c000000,
+        0x0000000038000000,
+        0x0000000030000000,
+        0x0000000020000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000002000000000,
+        0,
+        0x0000004000000000,
+        0,
+        0,
+        0,
+        0,
+        0x0000102000000000,
+        0,
+        0,
+        0x0000404000000000,
+        0,
+        0,
+        0,
+        0x0008102000000000,
+        0,
+        0,
+        0,
+        0x0040404000000000,
+        0,
+    ],
+    [
+        0,
+        0,
+        0,
+        0,
+        0x0000000000402000,
+        0,
+        0,
+        0x0000000000808000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000000400000,
+        0,
+        0x0000000000800000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x000000007e000000,
+        0x000000007c000000,
+        0x0000000078000000,
+        0x0000000070000000,
+        0x0000000060000000,
+        0x0000000040000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000004000000000,
+        0,
+        0x0000008000000000,
+        0,
+        0,
+        0,
+        0,
+        0x0000204000000000,
+        0,
+        0,
+        0x0000808000000000,
+        0,
+        0,
+        0,
+        0x0010204000000000,
+        0,
+        0,
+        0,
+        0x0080808000000000,
+    ],
+    [
+        0x0000000001010100,
+        0,
+        0,
+        0,
+        0x0000000002040800,
+        0,
+        0,
+        0,
+        0x0000000001010000,
+        0,
+        0,
+        0x0000000002040000,
+        0,
+        0,
+        0,
+        0,
+        0x0000000001000000,
+        0,
+        0x0000000002000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000200000000,
+        0x0000000600000000,
+        0x0000000e00000000,
+        0x0000001e00000000,
+        0x0000003e00000000,
+        0x0000007e00000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000010000000000,
+        0,
+        0x0000020000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0001010000000000,
+        0,
+        0,
+        0x0004020000000000,
+        0,
+        0,
+        0,
+        0,
+    ],
+    [
+        0,
+        0x0000000002020200,
+        0,
+        0,
+        0,
+        0x0000000004081000,
+        0,
+        0,
+        0,
+        0x0000000002020000,
+        0,
+        0,
+        0x0000000004080000,
+        0,
+        0,
+        0,
+        0,
+        0x0000000002000000,
+        0,
+        0x0000000004000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000400000000,
+        0x0000000c00000000,
+        0x0000001c00000000,
+        0x0000003c00000000,
+        0x0000007c00000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000020000000000,
+        0,
+        0x0000040000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0002020000000000,
+        0,
+        0,
+        0x0008040000000000,
+        0,
+        0,
+        0,
+    ],
+    [
+        0,
+        0,
+        0x0000000004040400,
+        0,
+        0,
+        0,
+        0x0000000008102000,
+        0,
+        0,
+        0,
+        0x0000000004040000,
+        0,
+        0,
+        0x0000000008100000,
+        0,
+        0,
+        0x0000000002000000,
+        0,
+        0x0000000004000000,
+        0,
+        0x0000000008000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000200000000,
+        0,
+        0,
+        0,
+        0x0000000800000000,
+        0x0000001800000000,
+        0x0000003800000000,
+        0x0000007800000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000020000000000,
+        0,
+        0x0000040000000000,
+        0,
+        0x0000080000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0004040000000000,
+        0,
+        0,
+        0x0010080000000000,
+        0,
+        0,
+    ],
+    [
+        0,
+        0,
+        0,
+        0x0000000008080800,
+        0,
+        0,
+        0,
+        0x0000000010204000,
+        0x0000000004020000,
+        0,
+        0,
+        0x0000000008080000,
+        0,
+        0,
+        0x0000000010200000,
+        0,
+        0,
+        0x0000000004000000,
+        0,
+        0x0000000008000000,
+        0,
+        0x0000000010000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000600000000,
+        0x0000000400000000,
+        0,
+        0,
+        0,
+        0x0000001000000000,
+        0x0000003000000000,
+        0x0000007000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000040000000000,
+        0,
+        0x0000080000000000,
+        0,
+        0x0000100000000000,
+        0,
+        0,
+        0x0002040000000000,
+        0,
+        0,
+        0x0008080000000000,
+        0,
+        0,
+        0x0020100000000000,
+        0,
+    ],
+    [
+        0x0000000008040200,
+        0,
+        0,
+        0,
+        0x0000000010101000,
+        0,
+        0,
+        0,
+        0,
+        0x0000000008040000,
+        0,
+        0,
+        0x0000000010100000,
+        0,
+        0,
+        0x0000000020400000,
+        0,
+        0,
+        0x0000000008000000,
+        0,
+        0x0000000010000000,
+        0,
+        0x0000000020000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000e00000000,
+        0x0000000c00000000,
+        0x0000000800000000,
+        0,
+        0,
+        0,
+        0x0000002000000000,
+        0x0000006000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000080000000000,
+        0,
+        0x0000100000000000,
+        0,
+        0x0000200000000000,
+        0,
+        0,
+        0x0004080000000000,
+        0,
+        0,
+        0x0010100000000000,
+        0,
+        0,
+        0x0040200000000000,
+    ],
+    [
+        0,
+        0x0000000010080400,
+        0,
+        0,
+        0,
+        0x0000000020202000,
+        0,
+        0,
+        0,
+        0,
+        0x0000000010080000,
+        0,
+        0,
+        0x0000000020200000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000010000000,
+        0,
+        0x0000000020000000,
+        0,
+        0x0000000040000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000001e00000000,
+        0x0000001c00000000,
+        0x0000001800000000,
+        0x0000001000000000,
+        0,
+        0,
+        0,
+        0x0000004000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000100000000000,
+        0,
+        0x0000200000000000,
+        0,
+        0x0000400000000000,
+        0,
+        0,
+        0x0008100000000000,
+        0,
+        0,
+        0x0020200000000000,
+        0,
+        0,
+    ],
+    [
+        0,
+        0,
+        0x0000000020100800,
+        0,
+        0,
+        0,
+        0x0000000040404000,
+        0,
+        0,
+        0,
+        0,
+        0x0000000020100000,
+        0,
+        0,
+        0x0000000040400000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000020000000,
+        0,
+        0x0000000040000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000003e00000000,
+        0x0000003c00000000,
+        0x0000003800000000,
+        0x0000003000000000,
+        0x0000002000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000200000000000,
+        0,
+        0x0000400000000000,
+        0,
+        0,
+        0,
+        0,
+        0x0010200000000000,
+        0,
+        0,
+        0x0040400000000000,
+        0,
+    ],
+    [
+        0,
+        0,
+        0,
+        0x0000000040201000,
+        0,
+        0,
+        0,
+        0x0000000080808000,
+        0,
+        0,
+        0,
+        0,
+        0x0000000040200000,
+        0,
+        0,
+        0x0000000080800000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000040000000,
+        0,
+        0x0000000080000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000007e00000000,
+        0x0000007c00000000,
+        0x0000007800000000,
+        0x0000007000000000,
+        0x0000006000000000,
+        0x0000004000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000400000000000,
+        0,
+        0x0000800000000000,
+        0,
+        0,
+        0,
+        0,
+        0x0020400000000000,
+        0,
+        0,
+        0x0080800000000000,
+    ],
+    [
+        0x0000000101010100,
+        0,
+        0,
+        0,
+        0,
+        0x0000000204081000,
+        0,
+        0,
+        0x0000000101010000,
+        0,
+        0,
+        0,
+        0x0000000204080000,
+        0,
+        0,
+        0,
+        0x0000000101000000,
+        0,
+        0,
+        0x0000000204000000,
+        0,
+        0,
+        0,
+        0,
+        0x0000000100000000,
+        0,
+        0x0000000200000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000020000000000,
+        0x0000060000000000,
+        0x00000e0000000000,
+        0x00001e0000000000,
+        0x00003e0000000000,
+        0x00007e0000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0001000000000000,
+        0,
+        0x0002000000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+    ],
+    [
+        0,
+        0x0000000202020200,
+        0,
+        0,
+        0,
+        0,
+        0x0000000408102000,
+        0,
+        0,
+        0x0000000202020000,
+        0,
+        0,
+        0,
+        0x0000000408100000,
+        0,
+        0,
+        0,
+        0x0000000202000000,
+        0,
+        0,
+        0x0000000408000000,
+        0,
+        0,
+        0,
+        0,
+        0x0000000200000000,
+        0,
+        0x0000000400000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000040000000000,
+        0x00000c0000000000,
+        0x00001c0000000000,
+        0x00003c0000000000,
+        0x00007c0000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0002000000000000,
+        0,
+        0x0004000000000000,
+        0,
+        0,
+        0,
+        0,
+    ],
+    [
+        0,
+        0,
+        0x0000000404040400,
+        0,
+        0,
+        0,
+        0,
+        0x0000000810204000,
+        0,
+        0,
+        0x0000000404040000,
+        0,
+        0,
+        0,
+        0x0000000810200000,
+        0,
+        0,
+        0,
+        0x0000000404000000,
+        0,
+        0,
+        0x0000000810000000,
+        0,
+        0,
+        0x0000000200000000,
+        0,
+        0x0000000400000000,
+        0,
+        0x0000000800000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000020000000000,
+        0,
+        0,
+        0,
+        0x0000080000000000,
+        0x0000180000000000,
+        0x0000380000000000,
+        0x0000780000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0002000000000000,
+        0,
+        0x0004000000000000,
+        0,
+        0x0008000000000000,
+        0,
+        0,
+        0,
+    ],
+    [
+        0,
+        0,
+        0,
+        0x0000000808080800,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000000808080000,
+        0,
+        0,
+        0,
+        0x0000001020400000,
+        0x0000000402000000,
+        0,
+        0,
+        0x0000000808000000,
+        0,
+        0,
+        0x0000001020000000,
+        0,
+        0,
+        0x0000000400000000,
+        0,
+        0x0000000800000000,
+        0,
+        0x0000001000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000060000000000,
+        0x0000040000000000,
+        0,
+        0,
+        0,
+        0x0000100000000000,
+        0x0000300000000000,
+        0x0000700000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0004000000000000,
+        0,
+        0x0008000000000000,
+        0,
+        0x0010000000000000,
+        0,
+        0,
+    ],
+    [
+        0,
+        0,
+        0,
+        0,
+        0x0000001010101000,
+        0,
+        0,
+        0,
+        0x0000000804020000,
+        0,
+        0,
+        0,
+        0x0000001010100000,
+        0,
+        0,
+        0,
+        0,
+        0x0000000804000000,
+        0,
+        0,
+        0x0000001010000000,
+        0,
+        0,
+        0x0000002040000000,
+        0,
+        0,
+        0x0000000800000000,
+        0,
+        0x0000001000000000,
+        0,
+        0x0000002000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x00000e0000000000,
+        0x00000c0000000000,
+        0x0000080000000000,
+        0,
+        0,
+        0,
+        0x0000200000000000,
+        0x0000600000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0008000000000000,
+        0,
+        0x0010000000000000,
+        0,
+        0x0020000000000000,
+        0,
+    ],
+    [
+        0x0000001008040200,
+        0,
+        0,
+        0,
+        0,
+        0x0000002020202000,
+        0,
+        0,
+        0,
+        0x0000001008040000,
+        0,
+        0,
+        0,
+        0x0000002020200000,
+        0,
+        0,
+        0,
+        0,
+        0x0000001008000000,
+        0,
+        0,
+        0x0000002020000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000001000000000,
+        0,
+        0x0000002000000000,
+        0,
+        0x0000004000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x00001e0000000000,
+        0x00001c0000000000,
+        0x0000180000000000,
+        0x0000100000000000,
+        0,
+        0,
+        0,
+        0x0000400000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0010000000000000,
+        0,
+        0x0020000000000000,
+        0,
+        0x0040000000000000,
+    ],
+    [
+        0,
+        0x0000002010080400,
+        0,
+        0,
+        0,
+        0,
+        0x0000004040404000,
+        0,
+        0,
+        0,
+        0x0000002010080000,
+        0,
+        0,
+        0,
+        0x0000004040400000,
+        0,
+        0,
+        0,
+        0,
+        0x0000002010000000,
+        0,
+        0,
+        0x0000004040000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000002000000000,
+        0,
+        0x0000004000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x00003e0000000000,
+        0x00003c0000000000,
+        0x0000380000000000,
+        0x0000300000000000,
+        0x0000200000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0020000000000000,
+        0,
+        0x0040000000000000,
+        0,
+    ],
+    [
+        0,
+        0,
+        0x0000004020100800,
+        0,
+        0,
+        0,
+        0,
+        0x0000008080808000,
+        0,
+        0,
+        0,
+        0x0000004020100000,
+        0,
+        0,
+        0,
+        0x0000008080800000,
+        0,
+        0,
+        0,
+        0,
+        0x0000004020000000,
+        0,
+        0,
+        0x0000008080000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000004000000000,
+        0,
+        0x0000008000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x00007e0000000000,
+        0x00007c0000000000,
+        0x0000780000000000,
+        0x0000700000000000,
+        0x0000600000000000,
+        0x0000400000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0040000000000000,
+        0,
+        0x0080000000000000,
+    ],
+    [
+        0x0000010101010100,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000020408102000,
+        0,
+        0x0000010101010000,
+        0,
+        0,
+        0,
+        0,
+        0x0000020408100000,
+        0,
+        0,
+        0x0000010101000000,
+        0,
+        0,
+        0,
+        0x0000020408000000,
+        0,
+        0,
+        0,
+        0x0000010100000000,
+        0,
+        0,
+        0x0000020400000000,
+        0,
+        0,
+        0,
+        0,
+        0x0000010000000000,
+        0,
+        0x0000020000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0002000000000000,
+        0x0006000000000000,
+        0x000e000000000000,
+        0x001e000000000000,
+        0x003e000000000000,
+        0x007e000000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+    ],
+    [
+        0,
+        0x0000020202020200,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000040810204000,
+        0,
+        0x0000020202020000,
+        0,
+        0,
+        0,
+        0,
+        0x0000040810200000,
+        0,
+        0,
+        0x0000020202000000,
+        0,
+        0,
+        0,
+        0x0000040810000000,
+        0,
+        0,
+        0,
+        0x0000020200000000,
+        0,
+        0,
+        0x0000040800000000,
+        0,
+        0,
+        0,
+        0,
+        0x0000020000000000,
+        0,
+        0x0000040000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0004000000000000,
+        0x000c000000000000,
+        0x001c000000000000,
+        0x003c000000000000,
+        0x007c000000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+    ],
+    [
+        0,
+        0,
+        0x0000040404040400,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000040404040000,
+        0,
+        0,
+        0,
+        0,
+        0x0000081020400000,
+        0,
+        0,
+        0x0000040404000000,
+        0,
+        0,
+        0,
+        0x0000081020000000,
+        0,
+        0,
+        0,
+        0x0000040400000000,
+        0,
+        0,
+        0x0000081000000000,
+        0,
+        0,
+        0x0000020000000000,
+        0,
+        0x0000040000000000,
+        0,
+        0x0000080000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0002000000000000,
+        0,
+        0,
+        0,
+        0x0008000000000000,
+        0x0018000000000000,
+        0x0038000000000000,
+        0x0078000000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+    ],
+    [
+        0,
+        0,
+        0,
+        0x0000080808080800,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000080808080000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000080808000000,
+        0,
+        0,
+        0,
+        0x0000102040000000,
+        0x0000040200000000,
+        0,
+        0,
+        0x0000080800000000,
+        0,
+        0,
+        0x0000102000000000,
+        0,
+        0,
+        0x0000040000000000,
+        0,
+        0x0000080000000000,
+        0,
+        0x0000100000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0006000000000000,
+        0x0004000000000000,
+        0,
+        0,
+        0,
+        0x0010000000000000,
+        0x0030000000000000,
+        0x0070000000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+    ],
+    [
+        0,
+        0,
+        0,
+        0,
+        0x0000101010101000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000101010100000,
+        0,
+        0,
+        0,
+        0x0000080402000000,
+        0,
+        0,
+        0,
+        0x0000101010000000,
+        0,
+        0,
+        0,
+        0,
+        0x0000080400000000,
+        0,
+        0,
+        0x0000101000000000,
+        0,
+        0,
+        0x0000204000000000,
+        0,
+        0,
+        0x0000080000000000,
+        0,
+        0x0000100000000000,
+        0,
+        0x0000200000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x000e000000000000,
+        0x000c000000000000,
+        0x0008000000000000,
+        0,
+        0,
+        0,
+        0x0020000000000000,
+        0x0060000000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+    ],
+    [
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000202020202000,
+        0,
+        0,
+        0x0000100804020000,
+        0,
+        0,
+        0,
+        0,
+        0x0000202020200000,
+        0,
+        0,
+        0,
+        0x0000100804000000,
+        0,
+        0,
+        0,
+        0x0000202020000000,
+        0,
+        0,
+        0,
+        0,
+        0x0000100800000000,
+        0,
+        0,
+        0x0000202000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000100000000000,
+        0,
+        0x0000200000000000,
+        0,
+        0x0000400000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x001e000000000000,
+        0x001c000000000000,
+        0x0018000000000000,
+        0x0010000000000000,
+        0,
+        0,
+        0,
+        0x0040000000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+    ],
+    [
+        0x0000201008040200,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000404040404000,
+        0,
+        0,
+        0x0000201008040000,
+        0,
+        0,
+        0,
+        0,
+        0x0000404040400000,
+        0,
+        0,
+        0,
+        0x0000201008000000,
+        0,
+        0,
+        0,
+        0x0000404040000000,
+        0,
+        0,
+        0,
+        0,
+        0x0000201000000000,
+        0,
+        0,
+        0x0000404000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000200000000000,
+        0,
+        0x0000400000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x003e000000000000,
+        0x003c000000000000,
+        0x0038000000000000,
+        0x0030000000000000,
+        0x0020000000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+    ],
+    [
+        0,
+        0x0000402010080400,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000808080808000,
+        0,
+        0,
+        0x0000402010080000,
+        0,
+        0,
+        0,
+        0,
+        0x0000808080800000,
+        0,
+        0,
+        0,
+        0x0000402010000000,
+        0,
+        0,
+        0,
+        0x0000808080000000,
+        0,
+        0,
+        0,
+        0,
+        0x0000402000000000,
+        0,
+        0,
+        0x0000808000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0000400000000000,
+        0,
+        0x0000800000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x007e000000000000,
+        0x007c000000000000,
+        0x0078000000000000,
+        0x0070000000000000,
+        0x0060000000000000,
+        0x0040000000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+    ],
+    [
+        0x0001010101010100,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0002040810204000,
+        0x0001010101010000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0002040810200000,
+        0,
+        0x0001010101000000,
+        0,
+        0,
+        0,
+        0,
+        0x0002040810000000,
+        0,
+        0,
+        0x0001010100000000,
+        0,
+        0,
+        0,
+        0x0002040800000000,
+        0,
+        0,
+        0,
+        0x0001010000000000,
+        0,
+        0,
+        0x0002040000000000,
+        0,
+        0,
+        0,
+        0,
+        0x0001000000000000,
+        0,
+        0x0002000000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0200000000000000,
+        0x0600000000000000,
+        0x0e00000000000000,
+        0x1e00000000000000,
+        0x3e00000000000000,
+        0x7e00000000000000,
+    ],
+    [
+        0,
+        0x0002020202020200,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0002020202020000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0004081020400000,
+        0,
+        0x0002020202000000,
+        0,
+        0,
+        0,
+        0,
+        0x0004081020000000,
+        0,
+        0,
+        0x0002020200000000,
+        0,
+        0,
+        0,
+        0x0004081000000000,
+        0,
+        0,
+        0,
+        0x0002020000000000,
+        0,
+        0,
+        0x0004080000000000,
+        0,
+        0,
+        0,
+        0,
+        0x0002000000000000,
+        0,
+        0x0004000000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0400000000000000,
+        0x0c00000000000000,
+        0x1c00000000000000,
+        0x3c00000000000000,
+        0x7c00000000000000,
+    ],
+    [
+        0,
+        0,
+        0x0004040404040400,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0004040404040000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0004040404000000,
+        0,
+        0,
+        0,
+        0,
+        0x0008102040000000,
+        0,
+        0,
+        0x0004040400000000,
+        0,
+        0,
+        0,
+        0x0008102000000000,
+        0,
+        0,
+        0,
+        0x0004040000000000,
+        0,
+        0,
+        0x0008100000000000,
+        0,
+        0,
+        0x0002000000000000,
+        0,
+        0x0004000000000000,
+        0,
+        0x0008000000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0200000000000000,
+        0,
+        0,
+        0,
+        0x0800000000000000,
+        0x1800000000000000,
+        0x3800000000000000,
+        0x7800000000000000,
+    ],
+    [
+        0,
+        0,
+        0,
+        0x0008080808080800,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0008080808080000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0008080808000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0008080800000000,
+        0,
+        0,
+        0,
+        0x0010204000000000,
+        0x0004020000000000,
+        0,
+        0,
+        0x0008080000000000,
+        0,
+        0,
+        0x0010200000000000,
+        0,
+        0,
+        0x0004000000000000,
+        0,
+        0x0008000000000000,
+        0,
+        0x0010000000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0600000000000000,
+        0x0400000000000000,
+        0,
+        0,
+        0,
+        0x1000000000000000,
+        0x3000000000000000,
+        0x7000000000000000,
+    ],
+    [
+        0,
+        0,
+        0,
+        0,
+        0x0010101010101000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0010101010100000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0010101010000000,
+        0,
+        0,
+        0,
+        0x0008040200000000,
+        0,
+        0,
+        0,
+        0x0010101000000000,
+        0,
+        0,
+        0,
+        0,
+        0x0008040000000000,
+        0,
+        0,
+        0x0010100000000000,
+        0,
+        0,
+        0x0020400000000000,
+        0,
+        0,
+        0x0008000000000000,
+        0,
+        0x0010000000000000,
+        0,
+        0x0020000000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0e00000000000000,
+        0x0c00000000000000,
+        0x0800000000000000,
+        0,
+        0,
+        0,
+        0x2000000000000000,
+        0x6000000000000000,
+    ],
+    [
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0020202020202000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0020202020200000,
+        0,
+        0,
+        0x0010080402000000,
+        0,
+        0,
+        0,
+        0,
+        0x0020202020000000,
+        0,
+        0,
+        0,
+        0x0010080400000000,
+        0,
+        0,
+        0,
+        0x0020202000000000,
+        0,
+        0,
+        0,
+        0,
+        0x0010080000000000,
+        0,
+        0,
+        0x0020200000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0010000000000000,
+        0,
+        0x0020000000000000,
+        0,
+        0x0040000000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x1e00000000000000,
+        0x1c00000000000000,
+        0x1800000000000000,
+        0x1000000000000000,
+        0,
+        0,
+        0,
+        0x4000000000000000,
+    ],
+    [
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0040404040404000,
+        0,
+        0x0020100804020000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0040404040400000,
+        0,
+        0,
+        0x0020100804000000,
+        0,
+        0,
+        0,
+        0,
+        0x0040404040000000,
+        0,
+        0,
+        0,
+        0x0020100800000000,
+        0,
+        0,
+        0,
+        0x0040404000000000,
+        0,
+        0,
+        0,
+        0,
+        0x0020100000000000,
+        0,
+        0,
+        0x0040400000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0020000000000000,
+        0,
+        0x0040000000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x3e00000000000000,
+        0x3c00000000000000,
+        0x3800000000000000,
+        0x3000000000000000,
+        0x2000000000000000,
+        0,
+        0,
+        0,
+    ],
+    [
+        0x0040201008040200,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0080808080808000,
+        0,
+        0x0040201008040000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0080808080800000,
+        0,
+        0,
+        0x0040201008000000,
+        0,
+        0,
+        0,
+        0,
+        0x0080808080000000,
+        0,
+        0,
+        0,
+        0x0040201000000000,
+        0,
+        0,
+        0,
+        0x0080808000000000,
+        0,
+        0,
+        0,
+        0,
+        0x0040200000000000,
+        0,
+        0,
+        0x0080800000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x0040000000000000,
+        0,
+        0x0080000000000000,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0x7e00000000000000,
+        0x7c00000000000000,
+        0x7800000000000000,
+        0x7000000000000000,
+        0x6000000000000000,
+        0x4000000000000000,
+        0,
+        0,
+    ],
+];
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    fn bitboard_index_to_position(index: usize) -> (usize, usize) {
+        (index / 8, index % 8)
+    }
+
+    fn position_to_bitboard_index(rank: usize, file: usize) -> usize {
+        rank * 8 + file
+    }
+
+    fn format_index(index: usize) -> String {
+        let (rank, file) = bitboard_index_to_position(index);
+        format!("{}{}", (file + 97) as u8 as char, rank + 1)
+    }
+
+    fn format_bitboard(bitboard: Bitboard) -> String {
+        let mut result = String::new();
+        result += "  abcdefgh\n";
+        for rank in (0..8).rev() {
+            result += &format!("{} ", rank + 1);
+            for file in 0..8 {
+                let index = position_to_bitboard_index(rank, file);
+                let bit = (bitboard >> index) & 1;
+                result += &format!("{}", bit);
+            }
+            result += "\n";
+        }
+        result
+    }
+
+    #[test]
+    fn check_in_between_table() {
+        let inbetween_table = super::IN_BETWEEN_TABLE;
+        for source in 0..64 {
+            for target in 0..64 {
+                let (source_rank, source_file) = bitboard_index_to_position(source);
+                let (target_rank, target_file) = bitboard_index_to_position(target);
+                let in_between = inbetween_table[source][target];
+
+                let rank_distance = ((source_rank as i8) - (target_rank as i8)).abs();
+                let file_distance = ((source_file as i8) - (target_file as i8)).abs();
+
+                if rank_distance <= 1 && file_distance <= 1 {
+                    assert_eq!(in_between, 0);
+                    continue;
+                }
+
+                let same_rank = rank_distance == 0 && file_distance > 0;
+                let same_file = file_distance == 0 && rank_distance > 0;
+                let same_diagonal = rank_distance == file_distance;
+
+                if same_rank {
+                    let mut in_between_rank = 0;
+                    let file_increment = if source_file < target_file { 1i8 } else { -1i8 };
+                    let mut file = source_file;
+
+                    for _ in 1..file_distance {
+                        file = (file as i8 + file_increment) as usize;
+                        in_between_rank |=
+                            1 << position_to_bitboard_index(source_rank, file as usize);
+                    }
+                    assert_eq!(
+                        in_between,
+                        in_between_rank,
+                        "source: {}, target: {}, expected:\n{}\nactual:\n{}",
+                        format_index(source),
+                        format_index(target),
+                        format_bitboard(in_between_rank),
+                        format_bitboard(in_between),
+                    );
+                } else if same_file {
+                    let mut in_between_file = 0;
+                    let rank_increment = if source_rank < target_rank { 1i8 } else { -1i8 };
+                    let mut rank = source_rank;
+
+                    for _ in 1..rank_distance {
+                        rank = (rank as i8 + rank_increment) as usize;
+                        in_between_file |=
+                            1 << position_to_bitboard_index(rank as usize, source_file);
+                    }
+                    assert_eq!(in_between, in_between_file);
+                } else if same_diagonal {
+                    let mut in_between_diagonal = 0;
+                    let rank_increment = if source_rank < target_rank { 1i8 } else { -1i8 };
+                    let file_increment = if source_file < target_file { 1i8 } else { -1i8 };
+
+                    let mut rank = source_rank;
+                    let mut file = source_file;
+                    for _ in 1..rank_distance {
+                        rank = (rank as i8 + rank_increment) as usize;
+                        file = (file as i8 + file_increment) as usize;
+                        in_between_diagonal |= 1 << position_to_bitboard_index(rank, file);
+                    }
+
+                    assert_eq!(
+                        in_between,
+                        in_between_diagonal,
+                        "source: {}, target: {}, expected:\n{}\nactual:\n{}",
+                        format_index(source),
+                        format_index(target),
+                        format_bitboard(in_between_diagonal),
+                        format_bitboard(in_between),
+                    );
+                } else {
+                    assert_eq!(in_between, 0);
+                }
+            }
+        }
+    }
+}
